requirements.txt
pandas==2.1.4
numpy==1.26.3
matplotlib==3-8
seaborn==0.13
catboost==1.2
shap==0.44.1
scikit-learn==1.3.0
pyyaml==6.0.1